pyproject.toml
[project]
name = "short-bond-hunter"
version = "0.1.0"
description = ""
authors = [
    {name = "github-main-user",email = "kolya.skripn@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "tinkoff-investments (>=0.2.0b117,<0.3.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
